Fix from recent PR 80 to allow MDI Alias icon names
This update now checks for icon name, if that fails, THEN it checks through the aliases of icons. 

Looks like the icon list is checked A-Z, so some icons which have aliases higher than the actual name (e.g. water, alias = liquid-spot) gets picked before the actual name (icon liquid-spot has an alias of water) has chance to be picked.