Introducing 50% colors
Half-colors (50% black, 50% accent) can rendered in a nice checkerboard style if we use `dither=2` when uploading the image to the AP.
We now use dither style 2 as default, i.e., if passing `dither=False` in the service data, that way we can use the 50% colors and it does not seem to have any significant other effect.
When showing photos, passing `dither=True` in the service data will use the dither style 1, just as before, existing setups that rely on dithering continue to work as before.

We introduce the following colors:

  - `half_black`, `hb`, `gray`, `grey`, `g`
  - `half_accent`, `ha`
  - `half_red`, `hr`
  - `half_yellow`, `hy`