Handle errors when fetching all tags, add retries
Previous behaviour was to break cycle upon error that lead to incomplete or empty result. In turn that lead to deletion of some tags (and re-addition later). Now in case of error code makes up to 10 retries (total, for all pages) and then re-raises an exception instead of ignoring it.